move feature_style_processor to an implementation header file
This allows other applications to create custom instances of the feature_style_processor with their own template arguments without forcing freuquent recompiles in mapnik itself
